Replace Configurable with CircuitBreaker
