+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ require 'stoplight'
 require 'timecop'
 require_relative 'support/data_store/base'
 require_relative 'support/light/runnable'
-require_relative 'support/configurable'
+require_relative 'support/circuit_breaker'
 require_relative 'support/database_cleaner'
 require_relative 'support/exception_helpers'
