+++ spec/stoplight_spec.rb
@@ -37,4 +37,42 @@ RSpec.describe 'Stoplight' do
   it 'creates a stoplight' do
     expect(light).to eq(Stoplight::Builder.with(name: name))
   end
+
+  it 'is a class' do
+    expect(light).to be_kind_of(Stoplight::CircuitBreaker)
+  end
+
+  describe '#name' do
+    it 'reads the name' do
+      expect(light.name).to eql(name)
+    end
+  end
+
+  describe '#fallback' do
+    it 'is initially the default' do
+      expect(light.fallback).to eql(Stoplight::Default::FALLBACK)
+    end
+  end
+
+  describe '#error_handler' do
+    it 'it initially the default' do
+      expect(light.error_handler).to eql(Stoplight::Default::ERROR_HANDLER)
+    end
+  end
+
+  describe '#with_error_handler' do
+    it 'sets the error handler' do
+      error_handler = ->(_, _) {}
+      with_error_handler = light.with_error_handler(&error_handler)
+      expect(with_error_handler.error_handler).to eql(error_handler)
+    end
+  end
+
+  describe '#with_fallback' do
+    it 'sets the fallback' do
+      fallback = ->(_) {}
+      with_fallback = light.with_fallback(&fallback)
+      expect(with_fallback.fallback).to eql(fallback)
+    end
+  end
 end
